Read filing's contents using specified encoding
The `@encoding` instance variable on a `Filing` object is ignored in
methods such as `Filing#form_type`, which can lead to an `ArgumentError`
("invalid byte sequence in UTF-8"). The included test case demonstrates
such an error when we try to call `Filing#summary`.

This change takes `@encoding` into account when reading the filing from
disk, which avoids the `ArgumentError`.